Maven Surefire Global Variable
Added a maven.surefire.version variable so that it can be used in
all of the pom.xml files to keep them aligned for future. Updated
maven surefire plugin to latest version (3.0.0-M5)

Fix an issue with github actions for DesktopTests that was causing
an issue with execution. Reverted back to calling DesktopTests
without a specific function until we figure out why using